CDPD-88335. HDDS-13544. DN Decommission Fails When Other Datanodes Are Offline Due to Invalid Affinity Node in Ratis Replication (#8934)
(cherry picked from commit 0fda8b28859855090fd27665edf92d36e84b7f70)

 Conflicts:
	hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/container/replication/ReplicationManagerUtil.java
	hadoop-hdds/server-scm/src/test/java/org/apache/hadoop/hdds/scm/container/replication/TestReplicationManagerUtil.java
	hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/hdds/scm/container/replication/TestReplicationManagerIntegration.java

TestReplicationManagerIntegration was not present in this version, so I added the entire class but only included the tests introduced by this commit. The Mini ozone cluster in this version doesn't work as expected with configurations, so I also needed to specify heartbeat intervals in the builder in the setup:
```
    MiniOzoneCluster.Builder builder = MiniOzoneCluster.newBuilder(conf)
        .setNumDatanodes(DATANODE_COUNT)
        .setHbInterval(100)
        .setHbProcessorInterval(101);

```
Had to additionally add the class TestNodeUtil for utility methods.